+++ client/src/App.jsx
@@ -1,13 +1,18 @@
+import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import Header from './components/Header/Header'
-import Experience from './components/Experience/Experience'
+import WelcomeSection from './components/Welcome/WelcomeSection'
+import AboutMe from './components/AboutMe/Aboutme'
 import './appStyling.css'
 
 function App() {
   return (
-    <>
+    <BrowserRouter>
       <Header />
-      <Experience />
-    </>
+      <Routes>
+        <Route path="/" element={<WelcomeSection />} />
+        <Route path="/aboutme" element={<AboutMe />} />
+      </Routes>
+    </BrowserRouter>
   )
 }
