+++ client/src/components/AboutMe/Aboutme.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Aboutme() {
+  return (
+    <div>Aboutme</div>
+  )
+}
