+++ client/src/components/Header/headerStyling.css
@@ -6,8 +6,12 @@
 
 /* Whole header section */
 .main_navbar_div {
-    height: 100vh;
     background-color: #222831;
+}
+
+.main_welcome_section {
+    background-color: #222831;
+    height: 100vh;
 }
 
 /* Top section */
@@ -68,6 +72,7 @@
 
 /* Welcome Section */
 .welcome_section {
+    background-color: #222831;
     width: 50%;
     margin: auto 50% auto auto;
     color: #EEEEEE;
