+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
-    "react-icons": "^5.1.0"
+    "react-icons": "^5.1.0",
+    "react-router-dom": "^6.23.0"
   }
 }
